모두 연결된 버전
