+++ app.py
@@ -1,38 +1,162 @@
-import IP_setup
-import visitors_counter
-import ssl
+import jwt
+import datetime
+import hashlib
 
 from pymongo import MongoClient
-from flask import Flask, render_template
+from flask import Flask, render_template, jsonify, request, redirect, url_for
+from datetime import datetime, timedelta
 
-ssl._create_default_https_context = ssl._create_unverified_context
+import IP_setup
+import visitors_counter
 
 app = Flask(__name__)
+app.config["TEMPLATES_AUTO_RELOAD"] = True
+app.config['UPLOAD_FOLDER'] = "./static/profile_pics"
 
+SECRET_KEY = 'SPARTA'
+
+# client = MongoClient('mongodb://15.164.50.47', 27017, username="test", password="test")
 client = MongoClient('localhost', 27017)  # 개인테스트용
-db = client.dbsparta
+db = client.dbsparta_plus_week4
 
 
-@app.route('/')  # IP_setup.IP() = [coordinate, region_korean, city_korean]
-def home():  # IP_setup.weather() = [temperature, description, humidity, image]
-    ip_index = list()
-    weather_index = list()
+@app.route('/')
+def home():
+    token_receive = request.cookies.get('mytoken')
+    try:
+        payload = jwt.decode(token_receive, SECRET_KEY, algorithms=['HS256'])
+        ip_index = list()
+        weather_index = list()
 
-    for i in IP_setup.IP():
-        ip_index.append(i)
+        for i in IP_setup.IP():
+            ip_index.append(i)
 
-    for j in IP_setup.weather(ip_index[0][0], ip_index[0][0]):  # ip_index[2][0], ip_index[2][1] = lat, lon
-        weather_index.append(j)
+        for j in IP_setup.weather(ip_index[0][0], ip_index[0][0]):  # ip_index[2][0], ip_index[2][1] = lat, lon
+            weather_index.append(j)
 
-    visitors_counter.visitors()
-    print(ip_index)
-    print(weather_index)
+        visitors_counter.visitors()
+        print(ip_index)
+        print(weather_index)
 
-    return render_template('index.html', region=ip_index[1], city=ip_index[2],
-                           temperature=weather_index[0], description=weather_index[1],
-                           humidity=weather_index[2], image=weather_index[3],
-                           visitors_today=list(db.todayCounter.find({}, {'_id': False})),
-                           visitors_total=list(db.visitorCounter.find({}, {'_id': False})))
+        return render_template('index.html', region=ip_index[1], city=ip_index[2],
+                               temperature=weather_index[0], description=weather_index[1],
+                               humidity=weather_index[2], image=weather_index[3],
+                               visitors_today=list(db.todayCounter.find({}, {'_id': False})),
+                               visitors_total=list(db.visitorCounter.find({}, {'_id': False})))
+    except jwt.ExpiredSignatureError:
+        return redirect(url_for("login", msg="로그인 시간이 만료되었습니다."))
+    except jwt.exceptions.DecodeError:
+        return redirect(url_for("login", msg="로그인 정보가 존재하지 않습니다."))
+
+
+@app.route('/login')
+def login():
+    msg = request.args.get("msg")
+    return render_template('login.html', msg=msg)
+
+
+@app.route('/user/<username>')
+def user(username):
+    # 각 사용자의 프로필과 글을 모아볼 수 있는 공간
+    token_receive = request.cookies.get('mytoken')
+    try:
+        payload = jwt.decode(token_receive, SECRET_KEY, algorithms=['HS256'])
+        status = (username == payload["id"])  # 내 프로필이면 True, 다른 사람 프로필 페이지면 False
+
+        user_info = db.users.find_one({"username": username}, {"_id": False})
+        return render_template('user.html', user_info=user_info, status=status)
+    except (jwt.ExpiredSignatureError, jwt.exceptions.DecodeError):
+        return redirect(url_for("home"))
+
+
+@app.route('/sign_in', methods=['POST'])
+def sign_in():
+    # 로그인
+    username_receive = request.form['username_give']
+    password_receive = request.form['password_give']
+
+    pw_hash = hashlib.sha256(password_receive.encode('utf-8')).hexdigest()
+    result = db.users.find_one({'username': username_receive, 'password': pw_hash})
+
+    if result is not None:
+        payload = {
+         'id': username_receive,
+         'exp': datetime.utcnow() + timedelta(seconds=60 * 60 * 24)  # 로그인 24시간 유지
+        }
+        token = jwt.encode(payload, SECRET_KEY, algorithm='HS256').decode('utf-8')
+
+        return jsonify({'result': 'success', 'token': token})
+    # 찾지 못하면
+    else:
+        return jsonify({'result': 'fail', 'msg': '아이디/비밀번호가 일치하지 않습니다.'})
+
+
+@app.route('/sign_up/save', methods=['POST'])
+def sign_up():
+    username_receive = request.form['username_give']
+    password_receive = request.form['password_give']
+    password_hash = hashlib.sha256(password_receive.encode('utf-8')).hexdigest()
+    doc = {
+        "username": username_receive,                               # 아이디
+        "password": password_hash,                                  # 비밀번호
+        "profile_name": username_receive,                           # 프로필 이름 기본값은 아이디
+        "profile_pic": "",                                          # 프로필 사진 파일 이름
+        "profile_pic_real": "profile_pics/profile_placeholder.png", # 프로필 사진 기본 이미지
+        "profile_info": ""                                          # 프로필 한 마디
+    }
+    db.users.insert_one(doc)
+    return jsonify({'result': 'success'})
+
+
+@app.route('/sign_up/check_dup', methods=['POST'])
+def check_dup():
+    username_receive = request.form['username_give']
+    exists = bool(db.users.find_one({"username": username_receive}))
+    return jsonify({'result': 'success', 'exists': exists})
+
+
+@app.route('/update_profile', methods=['POST'])
+def save_img():
+    token_receive = request.cookies.get('mytoken')
+    try:
+        payload = jwt.decode(token_receive, SECRET_KEY, algorithms=['HS256'])
+        # 프로필 업데이트
+        return jsonify({"result": "success", 'msg': '프로필을 업데이트했습니다.'})
+    except (jwt.ExpiredSignatureError, jwt.exceptions.DecodeError):
+        return redirect(url_for("home"))
+
+
+@app.route('/posting', methods=['POST'])
+def posting():
+    token_receive = request.cookies.get('mytoken')
+    try:
+        payload = jwt.decode(token_receive, SECRET_KEY, algorithms=['HS256'])
+        # 포스팅하기
+        return jsonify({"result": "success", 'msg': '포스팅 성공'})
+    except (jwt.ExpiredSignatureError, jwt.exceptions.DecodeError):
+        return redirect(url_for("home"))
+
+
+@app.route("/get_posts", methods=['GET'])
+def get_posts():
+    token_receive = request.cookies.get('mytoken')
+    try:
+        payload = jwt.decode(token_receive, SECRET_KEY, algorithms=['HS256'])
+        # 포스팅 목록 받아오기
+        return jsonify({"result": "success", "msg": "포스팅을 가져왔습니다."})
+    except (jwt.ExpiredSignatureError, jwt.exceptions.DecodeError):
+        return redirect(url_for("home"))
+
+
+@app.route('/update_like', methods=['POST'])
+def update_like():
+    token_receive = request.cookies.get('mytoken')
+    try:
+        payload = jwt.decode(token_receive, SECRET_KEY, algorithms=['HS256'])
+        # 좋아요 수 변경
+        return jsonify({"result": "success", 'msg': 'updated'})
+    except (jwt.ExpiredSignatureError, jwt.exceptions.DecodeError):
+        return redirect(url_for("home"))
 
 
 if __name__ == '__main__':
