--- templates/index_juny.html
@@ -1,260 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <!-- Bulma CSS -->
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.9.1/css/bulma.min.css">
-    <!-- Font Awesome CSS -->
-    <link href="//maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet">
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-    <link href="https://fonts.googleapis.com/css2?family=Gamja+Flower&family=Jua&display=swap" rel="stylesheet">
-
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-    <link href="https://fonts.googleapis.com/css2?family=Black+Han+Sans&family=Do+Hyeon&family=Jua&display=swap"
-          rel="stylesheet">
-    <!-- JS -->
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-cookie/1.4.1/jquery.cookie.js"></script>
-
-    <title>항해 날씨 - 메인페이지</title>
-
-
-    <style>
-        * {
-            margin: 0;
-            padding: 0;
-            list-style: none;
-            color: #000;
-            box-sizing: border-box;
-        }
-
-        body {
-            background: #ffffff;
-
-        }
-
-        a {
-            display: block;
-            text-decoration: none;
-        }
-
-        img {
-            display: block;
-        }
-
-        .wrap {
-            width: 70%;
-            margin: 0 auto;
-        }
-
-        .main_bg {
-            width: 100%;
-            height: 550px;
-            background-image: url("../static/763246.jpg");
-            background-size: cover;
-            background-position: center;
-        }
-
-        .box {
-            width: 80%;
-            margin: 0 auto;
-            margin-top: -50px;
-        }
-
-        .main_bg h1 {
-            text-align: center;
-            font-size: 3em;
-        }
-
-        .message {
-            width: 30%;
-            float: left;
-
-        }
-
-        .media{
-          width: 80%;
-          margin: 100px auto;
-        }
-
-        .weather_info{
-          width: 20%;
-          height: 400px;
-          margin-top: 40px;
-          margin-right: 40px;
-          float: right;
-          display: block;
-        }
-        .weather_info p{
-          text-align: right;
-        }
-    </style>
-</head>
-<body>
-<div class="wrap">
-    <section class="hero is-primary is-small">
-  <div class="hero-body">
-    <p class="title">
-      항해 날씨
-    </p>
-    <p class="subtitle">
-      subtitle
-    </p>
-  </div>
-</section>
-    <div class="main_bg">
-        <article class="message">
-            <div class="message-body">
-                항해하기 좋은날이네요.
-            </div>
-        </article>
-
-        <div class="weather_info box">
-          <p>님의 위치는 <br> 시입니다.</p>
-          <p>현재 온도 <br> 99도</p>
-
-        </div>
-
-    </div>
-    <div class="box">
-        <nav class="level is-mobile">
-            <div class="level-item has-text-centered">
-                <div>
-                    <p class="heading">Tweets</p>
-                    <p class="title">3,456</p>
-                </div>
-            </div>
-            <div class="level-item has-text-centered">
-                <div>
-                    <p class="heading">Following</p>
-                    <p class="title">123</p>
-                </div>
-            </div>
-            <div class="level-item has-text-centered">
-                <div>
-                    <p class="heading">Followers</p>
-                    <p class="title">456K</p>
-                </div>
-            </div>
-            <div class="level-item has-text-centered">
-                <div>
-                    <p class="heading">Likes</p>
-                    <p class="title">789</p>
-                </div>
-            </div>
-        </nav>
-    </div>
-
-  <article class="media">
-  <figure class="media-left">
-    <p class="image is-64x64">
-      <img src="https://bulma.io/images/placeholders/128x128.png">
-    </p>
-  </figure>
-  <div class="media-content">
-    <div class="field">
-      <p class="control">
-        <textarea class="textarea" placeholder="Add a comment..."></textarea>
-      </p>
-    </div>
-    <nav class="level">
-      <div class="level-left">
-        <div class="level-item">
-          <a class="button is-info">Submit</a>
-        </div>
-      </div>
-      <div class="level-right">
-        <div class="level-item">
-          <label class="checkbox">
-            <input type="checkbox"> Press enter to submit
-          </label>
-        </div>
-      </div>
-    </nav>
-  </div>
-</article>
-
-
-  <article class="media">
-  <figure class="media-left">
-    <p class="image is-64x64">
-      <img src="https://bulma.io/images/placeholders/128x128.png">
-    </p>
-  </figure>
-  <div class="media-content">
-    <div class="content">
-      <p>
-        <strong>Barbara Middleton</strong>
-        <br>
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis porta eros lacus, nec ultricies elit blandit non. Suspendisse pellentesque mauris sit amet dolor blandit rutrum. Nunc in tempus turpis.
-        <br>
-        <small><a>Like</a> · <a>Reply</a> · 3 hrs</small>
-      </p>
-    </div>
-
-    <article class="media">
-      <figure class="media-left">
-        <p class="image is-48x48">
-          <img src="https://bulma.io/images/placeholders/96x96.png">
-        </p>
-      </figure>
-      <div class="media-content">
-        <div class="content">
-          <p>
-            <strong>Sean Brown</strong>
-            <br>
-            Donec sollicitudin urna eget eros malesuada sagittis. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. Aliquam blandit nisl a nulla sagittis, a lobortis leo feugiat.
-            <br>
-            <small><a>Like</a> · <a>Reply</a> · 2 hrs</small>
-          </p>
-        </div>
-
-        <article class="media">
-          Vivamus quis semper metus, non tincidunt dolor. Vivamus in mi eu lorem cursus ullamcorper sit amet nec massa.
-        </article>
-
-        <article class="media">
-          Morbi vitae diam et purus tincidunt porttitor vel vitae augue. Praesent malesuada metus sed pharetra euismod. Cras tellus odio, tincidunt iaculis diam non, porta aliquet tortor.
-        </article>
-      </div>
-    </article>
-
-    <article class="media">
-      <figure class="media-left">
-        <p class="image is-48x48">
-          <img src="https://bulma.io/images/placeholders/96x96.png">
-        </p>
-      </figure>
-      <div class="media-content">
-        <div class="content">
-          <p>
-            <strong>Kayli Eunice </strong>
-            <br>
-            Sed convallis scelerisque mauris, non pulvinar nunc mattis vel. Maecenas varius felis sit amet magna vestibulum euismod malesuada cursus libero. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia Curae; Phasellus lacinia non nisl id feugiat.
-            <br>
-            <small><a>Like</a> · <a>Reply</a> · 2 hrs</small>
-          </p>
-        </div>
-      </div>
-    </article>
-  </div>
-</article>
-
-
-
-    <footer class="footer">
-        <div class="content has-text-centered">
-            <p>
-                <strong>항해날씨</strong> by 32조<a href="#">푸터영역입니다.</a>. The source code is licensed
-                <a href="#">MIT</a>.
-            </p>
-        </div>
-    </footer>
-
-
-</div>
-
-</body>
-</html>
